+++ src/CollisionAvoidance/CollisionAvoidanceSettingsController.h
@@ -26,7 +26,7 @@ public:
    Q_PROPERTY(int subsampleAmount                  READ getSubsampleAmount WRITE setSubsampleAmount NOTIFY subsampleAmountChanged);
    Q_PROPERTY(int particles                        READ getParticles WRITE setParticles NOTIFY particlesChanged);
    Q_PROPERTY(int windowSize                       READ getWindowSize WRITE setWindowSize NOTIFY windowSizeChanged);
-   Q_PROPERTY(int inlierProportionThreshold        READ getInlierProportionThreshold WRITE setInlierProportionThreshold NOTIFY inlierProportionThresholdChanged);
+   Q_PROPERTY(double inlierProportionThreshold     READ getInlierProportionThreshold WRITE setInlierProportionThreshold NOTIFY inlierProportionThresholdChanged);
    Q_PROPERTY(int divergencePatchSize              READ getDivergencePatchSize WRITE setDivergencePatchSize NOTIFY divergencePatchSizeChanged);
    Q_PROPERTY(double divergenceThreshold           READ getDivergenceThreshold WRITE setDivergenceThreshold NOTIFY divergenceThresholdChanged);
 
@@ -37,7 +37,7 @@ public:
     int               getSubsampleAmount() const                  { return settings.getSubsampleAmount();           }
     int               getParticles() const                        { return settings.getParticles();                 }
     int               getWindowSize() const                       { return settings.getWindowSize();                }
-    int               getInlierProportionThreshold() const        { return settings.getInlierProportionThreshold(); }
+    double            getInlierProportionThreshold() const        { return settings.getInlierProportionThreshold(); }
     int               getDivergencePatchSize() const              { return settings.getDivergencePatchSize();       }
     double            getDivergenceThreshold() const              { return settings.getDivergenceThreshold();       }
 
@@ -49,7 +49,7 @@ signals:
     void subsampleAmountChanged(int value);
     void particlesChanged(int value);
     void windowSizeChanged(int value);
-    void inlierProportionThresholdChanged(int value);
+    void inlierProportionThresholdChanged(double value);
     void divergencePatchSizeChanged(int value);
     void divergenceThresholdChanged(double value);
 
@@ -61,7 +61,7 @@ public slots:
     void setSubsampleAmount(int subsampleAmount)                                  { settings.setSubsampleAmount(subsampleAmount); emit subsampleAmountChanged(subsampleAmount); settings.storeSettings();                     }
     void setParticles(int particles)                                              { settings.setParticles(particles); emit particlesChanged(particles); settings.storeSettings();                                  }
     void setWindowSize(int windowSize)                                            { settings.setWindowSize(windowSize); emit windowSizeChanged(windowSize); settings.storeSettings();                               }
-    void setInlierProportionThreshold(int inlierProportionThreshold)              { settings.setInlierProportionThreshold(inlierProportionThreshold); emit inlierProportionThresholdChanged(inlierProportionThreshold); settings.storeSettings(); }
+    void setInlierProportionThreshold(double inlierProportionThreshold)           { settings.setInlierProportionThreshold(inlierProportionThreshold); emit inlierProportionThresholdChanged(inlierProportionThreshold); settings.storeSettings(); }
     void setDivergencePatchSize(int divergencePatchSize)                          { settings.setDivergencePatchSize(divergencePatchSize); emit divergencePatchSizeChanged(divergencePatchSize); settings.storeSettings();             }
     void setDivergenceThreshold(double divergenceThreshold)                       { settings.setDivergenceThreshold(divergenceThreshold); emit divergenceThresholdChanged(divergenceThreshold); settings.storeSettings();        }
